The stream is created once in mrcam_init(), and destroyed in mrcam_free()
Instead of with each frame being captured. This makes the callback work. Using
the callback the frame processing is finished in the callback handler, and I
want to clear up the stream there. But that necessitates killing the
currently-running thread, which doesn't work.

I was using this global stream previously, but changed to one-stream-per-frame
with this commit:

  commit 95271a24c9ee8ac5513c747c26632ff813c59768
  Author: Dima Kogan <Dmitriy.Kogan@jpl.nasa.gov>
  Date:   Fri Mar 29 22:10:36 2024 -0700

    I arv_camera_create_stream() and arv_camera_set_acquisition_mode() each frame

    Previously I did this at init time only, but that doesn't work with my gige
    pleora frame grabber apperently

After this patch, I need to re-test that camera