pywrap no longer has image callback
This didn't work right, since the callback was being done from a different
thread. I tried to synchronize things, but all of the Python interpreter and
Fl_Gl_Image_Widget and FLTK were having trouble, and it looked like a losing
battle.

The new method uses a pipe to signal a frame being available. The main Python
loop can select() the fd of the pipe to know when an image is available, and it
can then do a nonblocking read of the image. FLTK can do that like this:

  Fl.add_fd( c.fd_image_ready, callback )

This is a good scheme. I can do all the heavy lifting in the non-main thread, so
the UI stays responsive. And when data is available, the main thread can swoop
in, and read it.