Added "acquisition-persistent" option to all the APIs
If enabled, this keeps a single acquisition going throughout the run of the
program. Prior to this I would do this for the CONTINUOUS acquisition mode. But
now I see that different cameras require different things. In particular, the
Emergent HR20000m requires the MULTI_FRAME acquisition mode and
acquire_persistent