image-saving is now done by a C thread
This doesn't use the python "multiprocess" module. That was doing mystery things
under the hood, and would produce scary-looking errors at shutdown. Sometimes.
And it was almost certainly less efficient, since I THINK it copies the whole
image to the child process.

Here I'm doing a simple thing. No copying. And it's a child thread not a
process.